Corrected some missing cross-language tests.
